refactor: move the helper funcs and constants from controllerserver.go into a helper file
